refactor(contracts): split validation contract types into dedicated files
Why:
- Validation 層で列挙体・エラー型・Reader 実装が同一ファイルに同居し、責務境界が曖昧だったため。

What Changed:
- ExpectationConstraint 系の値型・エラー型・スキーマ・値Readerを専用ファイルへ分離。
- JsonStringContractReader から PresenceRequirement / ReadErrorKind / ReadError 定義を分離。
- OperationContractReader から OperationObjectReadErrorKind 定義を分離。

Impact:
- 1ファイル1責務に近づき、型探索と変更影響範囲の把握が容易になった。既存の契約挙動は維持。
