+++ src/Ucli.Contracts/Ipc/Validation/JsonStringContractReader.cs
@@ -1,50 +1,6 @@
-using System;
 using System.Text.Json;
 
 namespace MackySoft.Ucli.Contracts.Ipc.Validation;
-
-/// <summary> Specifies presence requirements for JSON string properties. </summary>
-internal enum JsonStringPresenceRequirement
-{
-    /// <summary> Property must be present and must be a JSON string. </summary>
-    Required,
-
-    /// <summary> Property may be absent, but must be a JSON string when specified. </summary>
-    OptionalStrict,
-
-    /// <summary> Property may be absent or non-string, and non-string values are treated as unspecified. </summary>
-    OptionalLoose,
-}
-
-/// <summary> Defines machine-readable error kinds for JSON string-property reads. </summary>
-internal enum JsonStringReadErrorKind
-{
-    /// <summary> No error. </summary>
-    None = 0,
-
-    /// <summary> Property is required but missing. </summary>
-    Missing,
-
-    /// <summary> Property exists but is not a JSON string. </summary>
-    TypeMismatch,
-
-    /// <summary> Property string is empty or whitespace only. </summary>
-    EmptyOrWhitespace,
-
-    /// <summary> Property string contains leading or trailing whitespace. </summary>
-    OuterWhitespace,
-}
-
-/// <summary> Represents one JSON string-property read error. </summary>
-/// <param name="Kind"> The machine-readable error kind. </param>
-/// <param name="PropertyName"> The property name that failed contract validation. </param>
-internal readonly record struct JsonStringReadError (
-    JsonStringReadErrorKind Kind,
-    string PropertyName)
-{
-    /// <summary> Gets an empty error value that indicates success. </summary>
-    public static JsonStringReadError None => new(JsonStringReadErrorKind.None, string.Empty);
-}
 
 /// <summary> Provides reusable contract readers for JSON string properties. </summary>
 internal static class JsonStringContractReader
